Sync of 21/1/24 12:31

+++ src/main/java/me/elordenador/megajar/tema6/contarletras/ContarLetras.java
@@ -46,7 +46,7 @@ public class ContarLetras {
             // 100 - max
             // n - x
 
-            double percentage = n * 100 / contadorMax;
+            double percentage = n * (double) 100 / (double) contadorMax;
             System.out.printf("%-10s %-25d %15.2f %%%n", key, n, percentage);
             //System.out.printf("%s                 %d                 %.2f%n    %%", key, n, percentage);
             //System.out.println("  " + key +".             "+n+"             " + percentage + " %");
